refactor(code-styleguide): update docs & admin-ui structure to match the pattern
