+++ packages/admin-ui/src/abbr/abbr.style.ts
@@ -1,0 +1,6 @@
+import { style } from '@vtex/admin-ui-core'
+
+export const base = style({
+  textDecoration: 'none',
+  text: '$body',
+})
