+++ packages/admin-ui/src/center/stories/center.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import type { Meta, Story } from '@storybook/react'
 import { tag } from '@vtex/admin-ui-react'
 
-import { Center } from './center'
+import { Center } from '../center'
 
 export default {
   title: 'admin-ui/center',
